requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 9
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 9
# NeuroHistoWSS/graph_research/loaders/utils_wsi.py: 5
# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 16
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 9
# NeuroHistoWSS/graph_research/utils/post_seg.py: 14
Pillow == 8.3.1

# NeuroHistoWSS/graph_research/test.py: 5
# NeuroHistoWSS/graph_research/train.py: 7
PyYAML == 5.4.1

# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 19
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 17
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 22
histomicstk == 1.0.7

# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 10,11
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 10,11
# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 9
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 10,11
# NeuroHistoWSS/graph_research/utils/misc.py: 9
matplotlib == 3.2.1

# NeuroHistoWSS/graph_research/utils/misc.py: 10
mlxtend == 0.17.2

# NeuroHistoWSS/graph_research/evaluators/base_eval.py: 8
# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 3
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 3
# NeuroHistoWSS/graph_research/loaders/__init__.py: 6
# NeuroHistoWSS/graph_research/loaders/utils_wsi.py: 2,3
# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 5
# NeuroHistoWSS/graph_research/models/pooling/seam.py: 1
# NeuroHistoWSS/graph_research/test.py: 7
# NeuroHistoWSS/graph_research/train.py: 8
# NeuroHistoWSS/graph_research/trainers/base_trainer.py: 8
# NeuroHistoWSS/graph_research/trainers/patch_enc_simclr.py: 5
# NeuroHistoWSS/graph_research/trainers/wsi_seam_utils.py: 1
# NeuroHistoWSS/graph_research/trainers/wsi_seg_trainer.py: 5
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 8
# NeuroHistoWSS/graph_research/utils/data_aug.py: 1
# NeuroHistoWSS/graph_research/utils/misc.py: 3
# NeuroHistoWSS/graph_research/utils/post_seg.py: 5
numpy == 1.19.2

# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 1
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 1
# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 3
# NeuroHistoWSS/graph_research/trainers/wsi_seam_utils.py: 4
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 3
opencv_python == 4.3.0.36

# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 8
pandas == 1.1.5

# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 5
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 5
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 13
pycm == 2.7

# NeuroHistoWSS/graph_research/trainers/wsi_seam_utils.py: 5,6
pydensecrf == 1.0rc3

# NeuroHistoWSS/graph_research/loaders/utils_wsi.py: 7,8
# NeuroHistoWSS/graph_research/utils/post_seg.py: 12
scikit_image == 0.17.2

# NeuroHistoWSS/graph_research/utils/misc.py: 12
# NeuroHistoWSS/graph_research/utils/post_seg.py: 9,10
scikit_learn == 0.23.0

# NeuroHistoWSS/graph_research/utils/post_seg.py: 4,8,13
scipy == 1.5.2

# NeuroHistoWSS/graph_research/utils/misc.py: 8
seaborn == 0.10.1

# NeuroHistoWSS/setup.py: 3
setuptools == 52.0.0.post20210125

# NeuroHistoWSS/graph_research/train.py: 14
tensorboardX == 2.2

# NeuroHistoWSS/graph_research/evaluators/base_eval.py: 2,3
# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 1,2,8
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 1,2,8
# NeuroHistoWSS/graph_research/loaders/__init__.py: 2
# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 14
# NeuroHistoWSS/graph_research/models/__init__.py: 2
# NeuroHistoWSS/graph_research/models/backbones/__init__.py: 5
# NeuroHistoWSS/graph_research/models/backbones/resnet.py: 1,2,3
# NeuroHistoWSS/graph_research/models/backbones/unet.py: 1,2
# NeuroHistoWSS/graph_research/models/pooling/classic.py: 1,2,3,4
# NeuroHistoWSS/graph_research/models/pooling/seam.py: 1,2,3
# NeuroHistoWSS/graph_research/test.py: 7
# NeuroHistoWSS/graph_research/train.py: 8
# NeuroHistoWSS/graph_research/trainers/base_trainer.py: 2,3
# NeuroHistoWSS/graph_research/trainers/patch_enc_simclr.py: 1,2,3,4
# NeuroHistoWSS/graph_research/trainers/wsi_seam_utils.py: 2,3
# NeuroHistoWSS/graph_research/trainers/wsi_seg_trainer.py: 1,2,3,4
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 1,4,5,6
# NeuroHistoWSS/graph_research/utils/data_aug.py: 2,3
# NeuroHistoWSS/graph_research/utils/misc.py: 2
torch == 1.9.0

# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 7
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 7
# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 13
# NeuroHistoWSS/graph_research/trainers/patch_enc_simclr.py: 8
# NeuroHistoWSS/graph_research/trainers/wsi_seg_trainer.py: 8
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 15
# NeuroHistoWSS/graph_research/utils/data_aug.py: 4
torchvision == 0.10.0

# NeuroHistoWSS/graph_research/evaluators/wsi_seg_eval.py: 6
# NeuroHistoWSS/graph_research/evaluators/wsi_ss_eval.py: 6
# NeuroHistoWSS/graph_research/loaders/wsi_datasets.py: 10
# NeuroHistoWSS/graph_research/trainers/patch_enc_simclr.py: 7
# NeuroHistoWSS/graph_research/trainers/wsi_seg_trainer.py: 7
# NeuroHistoWSS/graph_research/trainers/wsi_ss_trainer.py: 14
tqdm == 4.62.2
